+++ extensions/ql-vscode/test/vscode-tests/minimal-workspace/local-databases.test.ts
@@ -749,17 +749,14 @@ describe("local databases", () => {
         mockDbItem.databaseUri,
       );
 
-      expect(resolveDatabaseContentsSpy).toBeCalledTimes(1);
+      expect(resolveDatabaseContentsSpy).toBeCalledTimes(2);
     });
 
     it("should set the database as the currently selected one", async () => {
-      const makeSelected = true;
-
       await databaseManager.openDatabase(
         {} as ProgressCallback,
         {} as CancellationToken,
         mockDbItem.databaseUri,
-        makeSelected,
       );
 
       expect(setCurrentDatabaseItemSpy).toBeCalledTimes(1);